requirements.txt
pywebview
yfinance
pandas
python-dotenv
openai
alpaca-trade-api